+++ package.json
@@ -14,8 +14,10 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "-": "^0.0.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "sass": "^1.69.5"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
