+++ src/components/Button/Button.module.scss
@@ -3,6 +3,6 @@
   float: right;
   width: 75px;
   height: 75px;
-  top: -62px;
-  right: -45px;
+  top: -15px;
+  right: 50px;
 }
