+++ src/components/Icon/Icon.module.scss
