+++ src/components/SearchPanel/SearchPanel.module.scss
@@ -1,10 +1,10 @@
-.search-container {
+.container {
   width: 490px;
   display: block;
   margin: 0 auto;
 }
 
-input#search-bar {
+.search-bar {
   margin: 0 auto;
   width: 300px;
   height: 45px;
@@ -12,7 +12,6 @@ input#search-bar {
   font-size: 1rem;
   border: 1px solid #d0cfce;
   outline: none;
-
   border-radius: 20px;
   padding-left: 20px;
   margin-right: 10px;
