+++ src/components/SearchResult/SearchResult.module.scss
